noticket - code review updates
